Added class Stack, applied basic functions. Next: queue.
The HW assignment from bottom of stack.ts:
Make a queue. You can use a double-ended linked list. Queue has 2 functions: enqueue() and dequeue(). enqueue() adds an item to the end of the line. dequeue() takes an item from the front of the line. It's a checkout line at the store.